Replace miniHRID  master07-referencing.adoc
Mini HRID is undefined. Based on the previous section I assume the original meaning is compatible with ‘interface HRID’. Though it’s not clear to me wether that’s compatible with the next sentence that dropping the namespace means it’s the same namespace as the current archetype.

It could also refer to the archetype_hrid from aom2. But that already states it’s only the major version number. So the mini could be removed. But since this is clearly a reference (not a class attribute) I think aom2 is not very relevant here.